+++ react/components/ScrollAnimation.js
@@ -23,6 +23,7 @@ const withScrollAnimation = (
 
     state = {
       animation: '',
+      lastScrollPos: 0,
       trigger: false
     }
     
@@ -38,20 +39,26 @@ const withScrollAnimation = (
       return this.state.trigger 
     }
 
-    getAnimation = scroll => {
-      const shouldAnimate = scroll >= anchor && !reverse
-      if ( shouldAnimate ) this.setState({ trigger:true })
-      
-      return shouldAnimate
-        ? this.animation
-        : this.reverseAnimation
-    } 
+    animate(scroll) {
+      const scrollingUp = this.state.lastScrollPos > scroll
+      const scrollingDown = this.state.lastScrollPos < scroll
+
+      scrollingUp && this.setState({
+        animation: this.reverseAnimation,
+        lastScrollPos: scroll
+      })
+
+      scrollingDown && this.setState({
+        animation: this.animation,
+        lastScrollPos: scroll
+      })
+    }
 
     handleScroll = () => {
       const scroll = window.scrollY
-      this.setState({
-        animation: this.getAnimation(scroll)
-      })
+      const shouldAnimate = scroll >= anchor && !reverse
+      this.setState({ trigger: shouldAnimate })
+      shouldAnimate && this.animate(scroll)
     }
 
     render() {